feat: Separate repository layer from user service
